feat: Add endpoint to retrieve active users
This commit adds a new endpoint to the UserController class that allows retrieving a list of active users. The endpoint is accessible via a GET request to the "/active" path. The list of active users is obtained from the ChatService and returned as a response in the UserInfo DTO format.